requirements.txt
Flask==2.3.3
langchain_community==0.1.0
Pillow==9.4.0